+++ modules/officex.py
@@ -2,7 +2,7 @@ __G__ = "(G)bd249ce4"
 
 from ..logger.logger import verbose, verbose_flag, verbose_timeout
 from ..mics.funcs import get_words_multi_files
-from ..general.archive import check_packed_files, unpack_file
+from ..modules.archive import check_packed_files, unpack_file
 from re import sub
 from xml.dom.minidom import parseString
 from xml.etree.cElementTree import XML as cetXML