+++ services/online/malshare.py
@@ -19,11 +19,11 @@ class MalShare:
                 copyfileobj(r.raw, f)
             return True
 
-    @verbose(True,verbose_flag,5,"Getting hash details from MalShare")
+    @verbose(True,verbose_flag,verbose_timeout,"Getting hash details from MalShare")
     def get_hash_details(self,hash) -> str:
         return dumps(get("{}&action=details&hash={}".format(self.link,hash),verify=False).json(),indent=4)
 
-    @verbose(True,verbose_flag,5,"Getting hash details from MalShare")
+    @verbose(True,verbose_flag,verbose_timeout,"Getting hash details from MalShare")
     def get_file(self,hash) -> bool:
         return self.download_file("{}&action=getfile&hash={}".format(self.link,hash),hash)